+++ bin/invoke.js
