+++ index.js
@@ -1,6 +1,4 @@
-// const privateDirectory = require('./lib/privateDirectory');
-// const publicDirectory = require('./lib/publicDirectory');
-const PublicFileManager = require('./lib/PublicFileManager');
-const tui = require('./lib/tui');
+const PublicFileManager = require('./lib/shared/PublicFileManager');
+const tui = require('./lib/shared/tui/tui');
 
 module.exports = { PublicFileManager, tui };
