+++ lib/cli/.gitignore
