+++ lib/cli/README.md
@@ -1,0 +1,15 @@
+bakOsFs
+=======
+
+Cli to interact with global balloons of information
+
+[![oclif](https://img.shields.io/badge/cli-oclif-brightgreen.svg)](https://oclif.io)
+[![Version](https://img.shields.io/npm/v/bakOsFs.svg)](https://npmjs.org/package/bakOsFs)
+[![Downloads/week](https://img.shields.io/npm/dw/bakOsFs.svg)](https://npmjs.org/package/bakOsFs)
+[![License](https://img.shields.io/npm/l/bakOsFs.svg)](https://github.com/scmilee/filesystemprototype/bakOsFs/blob/master/package.json)
+
+<!-- toc -->
+# Usage
+<!-- usage -->
+# Commands
+<!-- commands -->
