+++ lib/cli/bin/run.cmd
