+++ lib/cli/src/commands/share.js
