+++ lib/cli/src/index.js
