+++ lib/cli/test/mocha.opts
