+++ lib/shared/PublicFileManager.js
