+++ lib/shared/privateDirectory.js
